[BPK-940] Refactor ios and android tests into common file
